Switch to orthographic camera projection and update test scene
Changed Camera to use orthographic projection instead of perspective, with dynamic resizing via Q/E keys. Updated TestScene to use 1280x720 resolution and enabled the point light component. Modified window creation to use WS_OVERLAPPEDWINDOW and commented out several window event handlers. Renderer now sets display mode size from parameters. TestMain now instantiates TestScene instead of PathFindingScene.

+++ Engine/Camera.h
@@ -5,8 +5,8 @@ class Camera : public Component
 {
 	float m_fov;
 
-	UINT m_screenWidth;
-	UINT m_screenHeight;
+	float m_screenWidth;
+	float m_screenHeight;
 	float m_aspectRatio;
 
 	float m_nearPlane;
